+++ src/_data/navigation.11tydata.js
@@ -1,0 +1,16 @@
+module.exports = {
+  items: [
+  	{
+  		title: "Home",
+  		slug: "."
+  	},
+  	{
+  		title: "About",
+  		slug: "/about"
+  	},
+  	{
+  		title: "Contact",
+  		slug: "/contact"
+  	}
+  ]
+}